+++ arith/mersenne31/src/m31x16/m31_avx512.rs
@@ -9,7 +9,6 @@ use std::{
 };
 
 use arith::{field_common, Field, SimdField};
-use ark_std::Zero;
 use ethnum::U256;
 use rand::{Rng, RngCore};
 use serdes::{ExpSerde, SerdeResult};
@@ -162,21 +161,6 @@ impl Field for AVXM31 {
                 )
             },
         }
-    }
-
-    fn exp(&self, exponent: u128) -> Self {
-        let mut e = exponent;
-        let mut res = Self::one();
-        let mut t = *self;
-        while !e.is_zero() {
-            let b = e & 1;
-            if b == 1 {
-                res *= t;
-            }
-            t = t * t;
-            e >>= 1;
-        }
-        res
     }
 
     #[inline(always)]
@@ -387,7 +371,6 @@ impl Mul<M31> for AVXM31 {
 impl Add<M31> for AVXM31 {
     type Output = AVXM31;
     #[inline(always)]
-    #[allow(clippy::op_ref)]
     fn add(self, rhs: M31) -> Self::Output {
         self + AVXM31::pack_full(rhs)
     }
