+++ arith/src/field.rs
@@ -1,3 +1,4 @@
+use ark_std::Zero;
 use ethnum::U256;
 use rand::RngCore;
 use serdes::ExpSerde;
@@ -118,7 +119,20 @@ pub trait Field:
     }
 
     /// Exp
-    fn exp(&self, exponent: u128) -> Self;
+    fn exp(&self, exponent: u128) -> Self {
+        let mut e = exponent;
+        let mut res = Self::one();
+        let mut t = *self;
+        while !e.is_zero() {
+            let b = e & 1;
+            if b == 1 {
+                res *= t;
+            }
+            t = t * t;
+            e >>= 1;
+        }
+        res
+    }
 
     /// find the inverse of the element; return None if not exist
     fn inv(&self) -> Option<Self>;
